Move version.ReleaseVersion to be linux-only.
