+++ version/osversion_nonlinux.go
@@ -1,0 +1,14 @@
+// Copyright 2015 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+// +build !linux
+
+package version
+
+// TODO(ericsnow) Refactor dependents so we can remove this for non-linux.
+
+// ReleaseVersion is a function that has no meaning except on linux.
+func ReleaseVersion() string {
+	// TODO(ericsnow) Panic?
+	return ""
+}
