+++ version/version_test.go
@@ -5,8 +5,6 @@ package version_test
 
 import (
 	"encoding/json"
-	"io/ioutil"
-	"path/filepath"
 	"runtime"
 	"strings"
 
@@ -323,72 +321,6 @@ func (*suite) TestParseMajorMinor(c *gc.C) {
 			c.Check(major, gc.Equals, test.expectMajor)
 			c.Check(minor, gc.Equals, test.expectMinor)
 		}
-	}
-}
-
-func (s *suite) TestUseFastLXC(c *gc.C) {
-	for i, test := range []struct {
-		message        string
-		releaseContent string
-		expected       string
-	}{{
-		message: "missing release file",
-	}, {
-		message:        "OS release file is missing ID",
-		releaseContent: "some junk\nand more junk",
-	}, {
-		message: "precise release",
-		releaseContent: `
-NAME="Ubuntu"
-VERSION="12.04 LTS, Precise"
-ID=ubuntu
-ID_LIKE=debian
-PRETTY_NAME="Ubuntu 12.04.3 LTS"
-VERSION_ID="12.04"
-`,
-		expected: "12.04",
-	}, {
-		message: "trusty release",
-		releaseContent: `
-NAME="Ubuntu"
-VERSION="14.04.1 LTS, Trusty Tahr"
-ID=ubuntu
-ID_LIKE=debian
-PRETTY_NAME="Ubuntu 14.04.1 LTS"
-VERSION_ID="14.04"
-`,
-		expected: "14.04",
-	}, {
-		message: "minimal trusty release",
-		releaseContent: `
-ID=ubuntu
-VERSION_ID="14.04"
-`,
-		expected: "14.04",
-	}, {
-		message: "minimal unstable unicorn",
-		releaseContent: `
-ID=ubuntu
-VERSION_ID="14.10"
-`,
-		expected: "14.10",
-	}, {
-		message: "minimal jaunty",
-		releaseContent: `
-ID=ubuntu
-VERSION_ID="9.10"
-`,
-		expected: "9.10",
-	}} {
-		c.Logf("%v: %v", i, test.message)
-		filename := filepath.Join(c.MkDir(), "os-release")
-		s.PatchValue(version.OSReleaseFile, filename)
-		if test.releaseContent != "" {
-			err := ioutil.WriteFile(filename, []byte(test.releaseContent+"\n"), 0644)
-			c.Assert(err, jc.ErrorIsNil)
-		}
-		value := version.ReleaseVersion()
-		c.Assert(value, gc.Equals, test.expected)
 	}
 }
